Don't manage the smartd service when ensure => absent
If the smartd package is not installed, puppet will attempt to stop
a service that doesn't exist. For this reason, the smartd service should
not be managed if ensure is set to absent or purged.
If smartd is getting uninstalled, the deb script will take care to stop
the service anyway.